face_service/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
python-multipart==0.0.20
numpy==2.2.5
Pillow==11.2.1
face-recognition==1.3.0
dlib==19.24.6
